Added option to make automatic cleanup optional.  We occasionally want to turn off cleanup when we want to automate load against a loadbalancer and then study the underlying vm without the system deleting it

+++ tests/loadbalancer_siege.py
@@ -124,8 +124,11 @@ class testLoadBalancerSiege(unittest.TestCase):
         ##########################
         if self.main_lb_id:
             self.lb_id = self.main_lb_id
-        self.logging.info("Deleting loadbalancer: %s" %self.lb_id)
-        result = self.driver.delete_lb(self.lb_id)
+        if self.args.cleanupoff:
+            self.logging.info("NOT deleting loadbalancer: %s per user-specified flag..." %self.lb_id)
+        else:
+            self.logging.info("Deleting loadbalancer: %s" %self.lb_id)
+            result = self.driver.delete_lb(self.lb_id)
